requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.1.0
brotlipy==0.7.0
cachetools==5.5.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.8
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.12.1
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib_1652035107515/work
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grad-cam==1.5.5
grpcio==1.70.0
huggingface-hub==0.28.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
imutils==0.5.4
iopath==0.1.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.11.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.0.3
pathlib==1.0.1
pathtools==0.1.2
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
pretrainedmodels==0.7.4
protobuf==4.25.6
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.1.4
pyre-extensions==0.0.32
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
safetensors==0.5.2
scikit-image==0.21.0
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
segmentation-models-pytorch==0.3.3
sentry-sdk==1.17.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.11.1
tabulate==0.9.0
taming-transformers==0.0.1
tb-nightly==2.14.0a20230808
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
test_tube==0.7.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.9.16
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.1
torch==2.4.1
torch-fidelity==0.3.0
torchaudio==0.13.0+cu117
torchcam==0.4.0
torchmetrics==1.0.3
torchstat==0.0.7
torchsummary==1.5.1
torchtnt==0.2.4
torchvision==0.19.1
tqdm==4.67.1
transformers==4.35.2
triton==3.0.0
ttach==0.0.3
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1666298941550/work
wandb==0.14.0
wcwidth==0.2.13
webrtcvad==2.0.10
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.16.2
